fixed cron jobs
